+++ changelog.md
@@ -13,6 +13,8 @@ Changes performed to the current version:
    * modified the protocol to publish/unpublish project - use token based authentication
    * updated OmenAPI to create a config file in the userdir
    * use config file in userdir for various things (authentication token, which config repository to be used)
+   * added --fast option for create and init commands to skip the question answering
+   * use stored user and email for create and init commands
 
 ### Version 0.2.2